+++ frontend/src/features/tweets/api/getTweets.ts
@@ -1,0 +1,7 @@
+import axios from 'axios';
+import { Tweets } from '../types';
+
+export const getTweets = async (): Promise<Tweets> => {
+	const response = await axios.get('/api/tweets');
+	return response.data;
+};
