+++ frontend/src/types/index.ts
@@ -1,21 +1,3 @@
-export interface Tweet {
-	id: number;
-	twitterUserId: bigint;
-	scheduledTimeUTC: Date | string;
-	content: string;
-	sent: boolean;
-}
-
-// Array of Tweets
-export type Tweets = Array<Tweet>;
-
-// Tweet sent from frontend and received by backend
-export interface HTTPTweet {
-	userId: string;
-	text: string;
-	scheduledTimeUTC?: string;
-}
-
 export interface Paper {
 	type: string;
 	title: string;
